pyproject.toml
[project]
name = "code-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.73.1",
    "groq>=1.2.0",
    "python-dotenv>=1.2.2",
]
